--- services/ui-src/src/measures/2021/AMRAD/questions/index.tsx
@@ -1,1 +1,0 @@
-export * from "./OptionalMeasureStratification";
